230722 수업 코드 최신화

+++ Engine_SOURCE/Lu_Time.cpp
@@ -24,7 +24,7 @@ namespace Lu
 	{
 		QueryPerformanceCounter(&m_CurFrequency);
 
-		double differnceFrequency = (m_CurFrequency.QuadPart - m_PrevFrequency.QuadPart);
+		double differnceFrequency = (double)(m_CurFrequency.QuadPart - m_PrevFrequency.QuadPart);
 
 		m_DeltaTime = (differnceFrequency / m_CpuFrequency.QuadPart);
